% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{calcular_objetivo_distancia}
\alias{calcular_objetivo_distancia}
\title{Method to Calculate the Distance Objective}
\usage{
calcular_objetivo_distancia(solucion)
}
\arguments{
\item{solucion}{Solution class object containing the plane equation}
}
\value{
Solution class object with the distance objective calculated
}
\description{
This method calculates the distance objective based on the plane equation.
}
