% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{calcular_objetivo_epsilon}
\alias{calcular_objetivo_epsilon}
\title{Method to calculate the epsilon objective}
\usage{
calcular_objetivo_epsilon(solucion)
}
\arguments{
\item{solucion}{Solucion class object}
}
\value{
Solucion class object with the epsilon objective calculated
}
\description{
Method to calculate the epsilon objective
}
