% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{calculate_distance_objective}
\alias{calculate_distance_objective}
\title{Method to Calculate the Distance Objective}
\usage{
calculate_distance_objective(solution)
}
\arguments{
\item{solution}{Solution class object containing the plane equation}
}
\value{
Solution class object with the distance objective calculated
}
\description{
This method calculates the distance objective based on the plane equation.
}
