% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Classification Metrics}
\usage{
calculate_metrics(solution, y_true, y_pred, decision_scores)
}
\arguments{
\item{solution}{A solution class object where the metrics will be stored}

\item{y_true}{A vector of true class labels (-1 or 1)}

\item{y_pred}{A vector of predicted class labels (-1 or 1)}

\item{decision_scores}{A numeric vector of decision scores (f(x) values)}
}
\value{
The input solution object with the following additional fields:
\item{conf_matrix}{Confusion matrix}
\item{accuracy}{Classification accuracy}
\item{precision}{Precision}
\item{recall}{Recall}
\item{f1}{F1 score}
\item{kappa}{Cohen's kappa statistic}
\item{auc}{Area Under the ROC Curve (AUC)}
}
\description{
This function calculates common classification metrics including accuracy, precision,
recall, F1-score, Cohen's kappa, AUC, and the confusion matrix. The results are stored
as fields in the provided solution object.
}
