% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{calculate_misclassified_objective}
\alias{calculate_misclassified_objective}
\title{Method to Calculate the FP/FN Objective}
\usage{
calculate_misclassified_objective(solution)
}
\arguments{
\item{solution}{Solution class object containing data and features for classification}
}
\value{
Solution class object with the FP/FN objective calculated
}
\description{
This method calculates the FP/FN objective by evaluating misclassified points.
}
