% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{check_clones}
\alias{check_clones}
\title{Check for Clones}
\usage{
check_clones(population, solution_eva)
}
\arguments{
\item{population}{Population}

\item{solution_eva}{Solution object to be evaluated for duplication}
}
\value{
1 if a clone is found, otherwise 0
}
\description{
This function checks if a solution already exists in the solution list to avoid duplicates.
}
