% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{comparar_soluciones}
\alias{comparar_soluciones}
\title{Method to Compare Solutions}
\usage{
comparar_soluciones(solucion1, solucion2)
}
\arguments{
\item{solucion1}{First solution object}

\item{solucion2}{Second solution object}
}
\value{
The solution object with better dominance or front
}
\description{
This function compares two solutions based on their front and dominance information.
}
