% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{compare_solutions}
\alias{compare_solutions}
\title{Method to Compare Solutions}
\usage{
compare_solutions(solution1, solution2)
}
\arguments{
\item{solution1}{First solution object}

\item{solution2}{Second solution object}
}
\value{
The solution object with better dominance or front
}
\description{
This function compares two solutions based on their front and dominance information.
}
