% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{comprobar_clones}
\alias{comprobar_clones}
\title{Check for Clones}
\usage{
comprobar_clones(poblacion, solucion_eva)
}
\arguments{
\item{poblacion}{Population}

\item{solucion_eva}{Solution object to be evaluated for duplication}
}
\value{
1 if a clone is found, otherwise 0
}
\description{
This function checks if a solution already exists in the solution list to avoid duplicates.
}
