% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{construct_planes}
\alias{construct_planes}
\title{Method to Construct Planes}
\usage{
construct_planes(solution)
}
\arguments{
\item{solution}{Solution class object containing data and features to construct planes}
}
\value{
Solution class object with constructed planes
}
\description{
Method to construct planes by calculating the coefficients based on features.
}
