% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{construir_planos}
\alias{construir_planos}
\title{Method to Construct Planes}
\usage{
construir_planos(solucion)
}
\arguments{
\item{solucion}{Solution class object containing data and features to construct planes}
}
\value{
Solution class object with constructed planes
}
\description{
Method to construct planes by calculating the coefficients based on features.
}
