% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{crossover_solutions}
\alias{crossover_solutions}
\title{Crossover Solutions}
\usage{
crossover_solutions(population, parent, mother, num)
}
\arguments{
\item{population}{Population class object}

\item{parent}{First parent solution}

\item{mother}{Second parent solution}

\item{num}{Unique identifier for the new solutions}
}
\value{
List of two new offspring solutions
}
\description{
This function performs crossover between two parent solutions to create offspring.
}
