% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{cruzar_soluciones}
\alias{cruzar_soluciones}
\title{Crossover Solutions}
\usage{
cruzar_soluciones(poblacion, padre, madre, num)
}
\arguments{
\item{poblacion}{Population class object}

\item{padre}{First parent solution}

\item{madre}{Second parent solution}

\item{num}{Unique identifier for the new solutions}
}
\value{
List of two new offspring solutions
}
\description{
This function performs crossover between two parent solutions to create offspring.
}
