% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{dominar}
\alias{dominar}
\title{Method to Determine Dominance}
\usage{
dominar(solucion1, solucion2)
}
\arguments{
\item{solucion1}{First solution object}

\item{solucion2}{Second solution object}
}
\value{
TRUE if solucion1 dominates solucion2, otherwise FALSE
}
\description{
This function determines if one solution dominates another based on objectives.
}
