% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{dominar2}
\alias{dominar2}
\title{Method to Determine Dominance with Three States}
\usage{
dominar2(solucion1, solucion2)
}
\arguments{
\item{solucion1}{First solution object}

\item{solucion2}{Second solution object}
}
\value{
0 if the solutions are equal, 1 if solucion1 dominates solucion2, 2 if solucion2 dominates solucion1
}
\description{
This function determines the dominance state between two solutions.
}
