% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{dominate}
\alias{dominate}
\title{Method to Determine Dominance}
\usage{
dominate(solution1, solution2)
}
\arguments{
\item{solution1}{First solution object}

\item{solution2}{Second solution object}
}
\value{
TRUE if solucion1 dominates solucion2, otherwise FALSE
}
\description{
This function determines if one solution dominates another based on objectives.
}
