% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{dominate2}
\alias{dominate2}
\title{Method to Determine Dominance with Three States}
\usage{
dominate2(solution1, solution2)
}
\arguments{
\item{solution1}{First solution object}

\item{solution2}{Second solution object}
}
\value{
0 if the solutions are equal, 1 if solution1 dominates solution2, 2 if solution2 dominates solution1
}
\description{
This function determines the dominance state between two solutions.
}
