% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMFeature.R
\name{draw_solution}
\alias{draw_solution}
\title{Draw the Pareto Front of an SVMFeature object}
\usage{
draw_solution(x, ...)
}
\arguments{
\item{x}{An object of class "SVMFeature".}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{ggplot} representing the selected solution.
The plot summarizes the strcture of characteristics of the solution
generated by the optimization procedure.
}
\description{
Draw the Pareto Front of an SVMFeature object
}
