% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{equal}
\alias{equal}
\title{Compare equality of solutions}
\usage{
equal(solucion, other)
}
\arguments{
\item{solucion}{First solution}

\item{other}{Second solution}
}
\value{
TRUE if the solutions are equal, FALSE otherwise
}
\description{
This function compares if two solutions are equal.
}
