% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{evaluate_solution}
\alias{evaluate_solution}
\title{Evaluate Solution}
\usage{
evaluate_solution(solution)
}
\arguments{
\item{solution}{Solution class object to be evaluated}
}
\value{
Solution class object with updated evaluation status and objectives
}
\description{
This function evaluates a solution by calculating its objectives.
}
