% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{fast_non_dominated_sort}
\alias{fast_non_dominated_sort}
\title{Fast Non-Dominated Sorting Algorithm}
\usage{
fast_non_dominated_sort(poblacion)
}
\arguments{
\item{poblacion}{Population class object}
}
\value{
Number of fronts found
}
\description{
This function implements the fast non-dominated sorting algorithm.
}
