% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{fnds}
\alias{fnds}
\title{Fast Non-Dominated Sorting (fnds)}
\usage{
fnds(poblacion)
}
\arguments{
\item{poblacion}{Population class object}
}
\value{
Updated Population class object with sorted fronts
}
\description{
This function performs the Fast Non-Dominated Sorting on a given population.
}
