% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{generar_poblacion_inicial}
\alias{generar_poblacion_inicial}
\title{Generate initial population}
\usage{
generar_poblacion_inicial(poblacion)
}
\arguments{
\item{poblacion}{Population class object}
}
\value{
Population class object updated with the generated initial population
}
\description{
This function generates the initial population.
}
