% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{generate_initial_population}
\alias{generate_initial_population}
\title{Generate initial population}
\usage{
generate_initial_population(population)
}
\arguments{
\item{population}{Population class object}
}
\value{
Population class object updated with the generated initial population
}
\description{
This function generates the initial population.
}
