% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{generate_random_solution}
\alias{generate_random_solution}
\title{Method to Generate a Random Solution}
\usage{
generate_random_solution(solution)
}
\arguments{
\item{solution}{Solution class object to modify}
}
\value{
Solution class object with a randomly generated solution
}
\description{
Method to generate a random solution by selecting random features and class vectors.
}
