% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{get_class_vector}
\alias{get_class_vector}
\title{Method to Obtain a Random Class Vector}
\usage{
get_class_vector(solution, clazz)
}
\arguments{
\item{solution}{Solution class object containing the dataset}

\item{clazz}{Class for which to obtain a vector}
}
\value{
Index of the random class vector
}
\description{
Method to obtain a random class vector from the dataset.
}
