% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{imprimir_poblacion}
\alias{imprimir_poblacion}
\title{Print population}
\usage{
imprimir_poblacion(poblacion)
}
\arguments{
\item{poblacion}{Population class object}
}
\value{
Prints the population in the console
}
\description{
This function prints the population.
}
