% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{is_valid_solution}
\alias{is_valid_solution}
\title{Method to assess if a solution is valid}
\usage{
is_valid_solution(solution)
}
\arguments{
\item{solution}{A solution object}
}
\value{
TRUE if solution is valid, otherwise FALSE
}
\description{
This function determines if the objectives of a given solution
  contain numerical values.
}
