% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{mutar_solucion}
\alias{mutar_solucion}
\title{Mutate Solution}
\usage{
mutar_solucion(solucion, poblacion)
}
\arguments{
\item{solucion}{Solution object to mutate}

\item{poblacion}{Population class object that contains mutation parameters}
}
\value{
Mutated solution object
}
\description{
This function mutates a solution based on mutation probabilities and coordinates.
}
