% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{mutar_vectores}
\alias{mutar_vectores}
\title{Method to Mutate Vectors}
\usage{
mutar_vectores(solucion)
}
\arguments{
\item{solucion}{Solucion class object to mutate}
}
\value{
Solucion class object with updated vectors
}
\description{
This function mutates the vectors in the solution by assigning new class vectors.
}
