% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{mutate_vectors}
\alias{mutate_vectors}
\title{Method to Mutate Vectors}
\usage{
mutate_vectors(solution, data, output)
}
\arguments{
\item{solution}{Solution class object to mutate}

\item{data}{A data.frame with the input features.}

\item{output}{A vector with class labels.}
}
\value{
Solution class object with updated vectors
}
\description{
This function mutates the vectors in the solution by assigning new class vectors.
}
