% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{nueva_poblacion}
\alias{nueva_poblacion}
\title{Create New Population}
\usage{
nueva_poblacion(poblacion)
}
\arguments{
\item{poblacion}{Population class object}
}
\value{
Updated Population class object with the new population generated
}
\description{
This function creates a new population through crossover.
}
