% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{obtener_vector_clase}
\alias{obtener_vector_clase}
\title{Method to Obtain a Random Class Vector}
\usage{
obtener_vector_clase(solucion, clase)
}
\arguments{
\item{solucion}{Solucion class object containing the dataset}

\item{clase}{Class for which to obtain a vector}
}
\value{
Index of the random class vector
}
\description{
Method to obtain a random class vector from the dataset.
}
