% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{print_population}
\alias{print_population}
\title{Print population}
\usage{
print_population(population)
}
\arguments{
\item{population}{Population class object}
}
\value{
Prints the population in the console
}
\description{
This function prints the population.
}
