% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{reduce_population}
\alias{reduce_population}
\title{Reduce Population Size}
\usage{
reduce_population(population)
}
\arguments{
\item{population}{Population class object to reduce}
}
\value{
Reduced Population class object
}
\description{
This function reduces the population size based on fronts and crowding distance.
}
