% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{reducir_poblacion}
\alias{reducir_poblacion}
\title{Reduce Population Size}
\usage{
reducir_poblacion(poblacion)
}
\arguments{
\item{poblacion}{Population class object to reduce}
}
\value{
Reduced Population class object
}
\description{
This function reduces the population size based on fronts and crowding distance.
}
