% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMFeature.R
\name{run.SVMFeature}
\alias{run.SVMFeature}
\title{Run the SVMFeature optimization process}
\usage{
run.SVMFeature(object)
}
\arguments{
\item{object}{An object of class "SVMFeature".}
}
\value{
The final population after the optimization process.
}
\description{
Run the SVMFeature optimization process
}
