% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{select_best_offspring}
\alias{select_best_offspring}
\title{Compares two offspring and selects the best one based on dominance relations.}
\usage{
select_best_offspring(child1, child2)
}
\arguments{
\item{child1}{The first offspring to compare.}

\item{child2}{The second offspring to compare.}
}
\value{
List\link{Solution}
A list with two solutions, where the first one is the better offspring (the winner), and the second one
is the inferior offspring (the loser).
}
\description{
This method uses the concept of dominance in multi-objective optimization to compare the two offspring.
If one offspring dominates the other, the dominant one is selected. If neither dominates the other,
the tie is broken randomly.
}
