% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{Solucion}
\alias{Solucion}
\title{Definition of the Solucion class}
\usage{
Solucion(num, data, costes, inputs, output, num_features)
}
\arguments{
\item{num}{Number of the solution}

\item{data}{Data set}

\item{costes}{Cost vector}

\item{inputs}{Names of input variables}

\item{output}{Name of the output variable}

\item{num_features}{Number of features}
}
\value{
Solucion class object
}
\description{
Definition of the Solucion class
}
