% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{Solution}
\alias{Solution}
\title{Definition of the Solution Class}
\usage{
Solution(num, data, costs, inputs, output, num_features, obj)
}
\arguments{
\item{num}{Solution number}

\item{data}{Data set}

\item{costs}{Cost vector}

\item{inputs}{Names of input variables}

\item{output}{Name of the output variable}

\item{num_features}{Number of features}

\item{obj}{Objective}
}
\value{
Object of the \code{Solution} class
}
\description{
Defines the \code{Solution} class, which represents an individual solution.
}
