% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMFeature.R
\docType{class}
\name{SVMFeature-class}
\alias{SVMFeature-class}
\alias{SVMFeature}
\title{SVMFeature Class}
\arguments{
\item{data}{DataFrame. The input data.}

\item{inputs}{Character vector. The names of the input features.}

\item{output}{Character. The name of the output variable.}

\item{costs}{Numeric vector. The costs associated with the features.}

\item{pop_size}{Numeric. The population size.}

\item{num_fea}{Numeric. The number of features.}

\item{n_iter}{Numeric. The number of iterations.}
}
\value{
An instance of the SVMFeature class.

An initialized SVMFeature object.
Run the SVMFeature optimization process

The final population after the optimization process.
}
\description{
This class implements a feature selection algorithm using Support Vector Machines (SVM).
It initializes with user-provided data, runs the optimization process, and prints the final population.
}

\examples{
data <- read.table("data/prueba.txt", sep = ";", header = TRUE)
inputs <- c('x1', 'x2', 'x3')
output <- 'y'
costs <- c(5, 27, 10)
pop_size <- 10
num_fea <- 2
svm_feature <- SVMFeature$new(archivo, inputs, output, costs, pop_size, num_fea, n_iter, max_time, mode)
svm_feature$run()
Initialize the SVMFeature object

}
