% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMFeature.R
\name{SVMFeature}
\alias{SVMFeature}
\title{SVMFeature Class (S3 Implementation)}
\usage{
SVMFeature(
  data,
  inputs,
  output,
  costs,
  pop_size,
  num_fea,
  n_iter = 10,
  max_time = 300,
  mode = "iters",
  objective = "distance-epsilon",
  num_obj = 2,
  clones = 0,
  p_mutation = 0.7,
  p_mut_ind = 0.4,
  p_mut_fea = 0.4,
  p_mut_coord = 0.2,
  mut_coord = 0
)
}
\arguments{
\item{data}{DataFrame. The input data.}

\item{inputs}{Character vector. The names of the input features.}

\item{output}{Character. The name of the output variable.}

\item{costs}{Numeric vector. The costs associated with the features.}

\item{pop_size}{Numeric. The population size.}

\item{num_fea}{Numeric. The number of features.}

\item{n_iter}{Numeric. The number of iterations (default: 10).}

\item{max_time}{Numeric. The maximum execution time (default: 300 seconds).}

\item{mode}{Character. "iters" for iteration-based execution, "time" for time-based execution.}

\item{objective}{Character. "distance-epsilon" or "confusion-matrix".}

\item{num_obj}{Numeric. Number of objectives (default: 2).}

\item{clones}{Numeric. Number of clones (default: 0).}

\item{p_mutation}{Numeric. Mutation probability (default: 0.7).}

\item{p_mut_ind}{Numeric. Individual mutation probability (default: 0.4).}

\item{p_mut_fea}{Numeric. Feature mutation probability (default: 0.4).}

\item{p_mut_coord}{Numeric. Mutation coordinates probability (default: 0.2).}

\item{mut_coord}{Numeric. Mutation coordinates (default: 0).}
}
\value{
An S3 object of class "SVMFeature".
}
\description{
Implements a feature selection algorithm using Support Vector Machines (SVM).
}
