% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{to_dict}
\alias{to_dict}
\title{Method to convert the solution to a dictionary}
\usage{
to_dict(solution)
}
\arguments{
\item{solution}{Solution class object}
}
\value{
List with the information of the solution
}
\description{
Method to convert the solution to a dictionary
}
