% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{torneo_elegir_padre}
\alias{torneo_elegir_padre}
\title{Tournament Selection for Parent}
\usage{
torneo_elegir_padre(poblacion)
}
\arguments{
\item{poblacion}{Population class object}
}
\value{
Selected parent solution
}
\description{
This function selects a parent solution from the population using tournament selection.
}
