% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{tournament_select_parent}
\alias{tournament_select_parent}
\title{Tournament Selection for Parent}
\usage{
tournament_select_parent(population)
}
\arguments{
\item{population}{Population class object}
}
\value{
Selected parent solution
}
\description{
This function selects a parent solution from the population using tournament selection.
}
