% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMFeature.R
\name{update_df_solutions.SVMFeature}
\alias{update_df_solutions.SVMFeature}
\title{Update df_solutions for an SVMFeature object}
\usage{
update_df_solutions.SVMFeature(object)
}
\arguments{
\item{object}{An object of class "SVMFeature".}
}
\value{
The updated SVMFeature object.
}
\description{
Update df_solutions for an SVMFeature object
}
