% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{calcular_objetivo_distancia}
\alias{calcular_objetivo_distancia}
\title{Method to calculate the distance objective}
\usage{
calcular_objetivo_distancia(solucion)
}
\arguments{
\item{solucion}{Solucion class object}
}
\value{
Solucion class object with the distance objective calculated
}
\description{
Method to calculate the distance objective
}
