% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{calcular_objetivo_epsilon}
\alias{calcular_objetivo_epsilon}
\title{Method to Calculate the Epsilon Objective}
\usage{
calcular_objetivo_epsilon(solucion)
}
\arguments{
\item{solucion}{Solucion class object containing data and features for classification}
}
\value{
Solucion class object with the epsilon objective calculated
}
\description{
This method calculates the epsilon objective by evaluating misclassified points.
}
