% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{comprobar_clones}
\alias{comprobar_clones}
\title{Check clones in the population}
\usage{
comprobar_clones(solucion_eva, lista_soluciones)
}
\arguments{
\item{solucion_eva}{Solution to evaluate}

\item{lista_soluciones}{List of solutions}
}
\value{
TRUE if there are clones, FALSE otherwise
}
\description{
This function checks if there are clones in the population.
}
