% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{construct_planes}
\alias{construct_planes}
\title{Method to Construct Planes}
\usage{
construct_planes(solution, data)
}
\arguments{
\item{solution}{Solution class object containing data and features to construct planes}

\item{data}{A data.frame with the input features}
}
\value{
Solution class object with constructed planes
}
\description{
Method to construct planes by calculating the coefficients based on features.
}
