% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{construir_planos}
\alias{construir_planos}
\title{Method to construct planes}
\usage{
construir_planos(solucion)
}
\arguments{
\item{solucion}{Solucion class object}
}
\value{
Solucion class object with constructed planes
}
\description{
Method to construct planes
}
