% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{crowding_distance}
\alias{crowding_distance}
\title{Crowding Distance Calculation}
\usage{
crowding_distance(poblacion, num_front)
}
\arguments{
\item{poblacion}{Population class object}

\item{num_front}{The front number to calculate the crowding distance for}
}
\value{
List of solutions with updated crowding distances
}
\description{
This function calculates the crowding distance for solutions in a specified front.
}
