% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{evaluar_solucion}
\alias{evaluar_solucion}
\title{Method to evaluate the solution}
\usage{
evaluar_solucion(solucion)
}
\arguments{
\item{solucion}{Solucion class object}
}
\value{
Solucion class object with the evaluation performed
}
\description{
Method to evaluate the solution
}
