% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{evaluar_solucion}
\alias{evaluar_solucion}
\title{Evaluate Solution}
\usage{
evaluar_solucion(solucion)
}
\arguments{
\item{solucion}{Solucion class object to be evaluated}
}
\value{
Solucion class object with updated evaluation status and objectives
}
\description{
This function evaluates a solution by calculating its objectives.
}
