% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{generar_solucion_aleatoria}
\alias{generar_solucion_aleatoria}
\title{Method to Generate a Random Solution}
\usage{
generar_solucion_aleatoria(solucion)
}
\arguments{
\item{solucion}{Solucion class object to modify}
}
\value{
Solucion class object with a randomly generated solution
}
\description{
Method to generate a random solution by selecting random features and class vectors.
}
