% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{mutate_solution}
\alias{mutate_solution}
\title{Mutate Solution}
\usage{
mutate_solution(solution, population)
}
\arguments{
\item{solution}{Solution object to mutate}

\item{population}{Population class object that contains mutation parameters}
}
\value{
Mutated solution object
}
\description{
This function mutates a solution based on mutation probabilities and coordinates.
}
