% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{mutate_vectors}
\alias{mutate_vectors}
\title{Method to Mutate Vectors}
\usage{
mutate_vectors(solution)
}
\arguments{
\item{solution}{Solution class object to mutate}
}
\value{
Solution class object with updated vectors
}
\description{
This function mutates the vectors in the solution by assigning new class vectors.
}
