% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{new_population}
\alias{new_population}
\title{Create New Population}
\usage{
new_population(population)
}
\arguments{
\item{population}{Population class object}
}
\value{
Updated Population class object with the new population generated
}
\description{
This function creates a new population through crossover.
}
