% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poblacion.R
\name{Poblacion}
\alias{Poblacion}
\title{Class Population}
\usage{
Poblacion(data, costes, tam_pob, inputs, output, num_features, clones = 0)
}
\arguments{
\item{data}{Dataset}

\item{inputs}{Input variables}

\item{output}{Output variable}

\item{num_features}{Number of features}

\item{clones}{Number of clones (default: 0)}

\item{costs}{Associated costs}

\item{population_size}{Population size}
}
\value{
Population class object
}
\description{
This function defines the Population class with its attributes and methods.
}
