% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{Population}
\alias{Population}
\title{Definition of the Population Class}
\usage{
Population(
  data,
  costs,
  pop_size,
  inputs,
  output,
  num_features,
  num_obj = 2,
  clones = 0,
  p_mutation = 0.7,
  p_mut_ind = 0.4,
  p_mut_fea = 0.4,
  p_mut_coord = 0.2,
  mut_coord = 0
)
}
\arguments{
\item{data}{Dataset}

\item{costs}{Associated costs}

\item{pop_size}{Population size}

\item{inputs}{Input variables}

\item{output}{Output variable}

\item{num_features}{Number of features}

\item{num_obj}{Number of objective (default: 2)}

\item{clones}{Number of clones (default: 0)}

\item{p_mutation}{Mutation probability (default: 0.7)}

\item{p_mut_ind}{Individual mutation probability (default: 0.4)}

\item{p_mut_fea}{Feature mutation probability (default: 0.4)}

\item{p_mut_coord}{Mutation coordinates (default: 0.2)}

\item{mut_coord}{Mutation coordinates (default: 0)}
}
\value{
Object of the `Population` class
}
\description{
This function defines the `Population` class with its attributes and methods.
}
