% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMFeature.R
\docType{class}
\name{SVMFeature-class}
\alias{SVMFeature-class}
\alias{SVMFeature}
\title{SVMFeature Class}
\arguments{
\item{archivo}{Character. The file path of the input data.}

\item{inputs}{Character vector. The names of the input features.}

\item{output}{Character. The name of the output variable.}

\item{costes}{Numeric vector. The costs associated with the features.}

\item{tam_pob}{Numeric. The population size.}

\item{num_fea}{Numeric. The number of features.}

\item{n_iter}{Numeric. The number of iterations.}
}
\value{
An instance of the SVMFeature class.

An initialized SVMFeature object.
Run the SVMFeature optimization process

The final population after the optimization process.
}
\description{
This class implements a feature selection algorithm using Support Vector Machines (SVM).
It initializes with user-provided data, runs the optimization process, and prints the final population.
}

\examples{
archivo <- "data/prueba.txt"
inputs <- c('x1', 'x2', 'x3')
output <- 'y'
costes <- c(5, 27, 10)
tam_pob <- 10
num_fea <- 2
svm_feature <- SVMFeature$new(archivo, inputs, output, costes, tam_pob, num_fea, n_iter, tiempo_max, modo)
svm_feature$run()
Initialize the SVMFeature object

}
