% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solucion.R
\name{to_dict}
\alias{to_dict}
\title{Method to convert the solution to a dictionary}
\usage{
to_dict(solucion)
}
\arguments{
\item{solucion}{Solucion class object}
}
\value{
List with the information of the solution
}
\description{
Method to convert the solution to a dictionary
}
